Move commands to a "commands" directory
