+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__/
 .system.lock
 .idea
 .DS_Store
+venv
