+++ src/commands/remove_archived_tasks.py
@@ -3,20 +3,18 @@ import logging
 from notion.client import NotionClient
 import yaml
 
-import command
+import commands.command as command
 import lockfile
-import schedules
 import schema
 import storage
 
-DONE_STATUS = [ schema.DONE_STATUS, schema.NOT_DONE_STATUS ]
 LOGGER = logging.getLogger(__name__)
 
-class ArchiveDoneTasks(command.Command):
+class RemoveArchivedTasks(command.Command):
 
     @staticmethod
     def name():
-        return "archive-done-tasks"
+        return "remove-archived-tasks"
 
     @staticmethod
     def description():
@@ -27,7 +25,6 @@ class ArchiveDoneTasks(command.Command):
         parser.add_argument("--period", required=False,  default=[], action="append", help="The period for which the upsert should happen. Defaults to all")
 
     def run(self, args):
-
         period_filter = frozenset(p.lower() for p in args.period) if len(args.period) > 0 else None
 
         workspace = storage.load_workspace()
@@ -36,24 +33,22 @@ class ArchiveDoneTasks(command.Command):
             tasks = yaml.safe_load(tasks_file)
 
         client = NotionClient(token_v2=workspace["token_v2"])
+        self._remove_archived_tasks(period_filter, client, tasks, args.dry_run)
 
-        self._archive_done_tasks(period_filter, client, workspace, tasks, args.dry_run)
-
-    def _archive_done_tasks(self, period_filter, client, workspace, tasks, dry_run):
-
+    def _remove_archived_tasks(self, period_filter, client, tasks, dry_run):
         system_lock = lockfile.load_lock_file()
         project_lock = system_lock["projects"][tasks["key"]]
         root_page = client.get_block(project_lock["inbox"]["root_page_id"])
         page = client.get_collection_view(project_lock["inbox"]["database_view_id"], collection=root_page.collection)
 
-        all_done_tasks = page.build_query().execute()
+        archived_tasks = page.build_query().execute()
 
-        for done_task in all_done_tasks:
-            if done_task.status not in DONE_STATUS:
+        for archived_task in archived_tasks:
+            if archived_task.status != schema.ARCHIVED_STATUS:
                 continue
-            if period_filter and (done_task.script_period.lower() not in period_filter):
-                LOGGER.info("Skipping '{name}' on account of period filtering".format(name=done_task.name))
+            if period_filter and (archived_task.script_period.lower() not in period_filter):
+                LOGGER.info("Skipping '{name}' on account of period filtering".format(name=archived_task.name))
                 continue
-            LOGGER.info("Archiving '{name}'".format(name=done_task.name))
+            LOGGER.info("Removing '{name}'".format(name=archived_task.name))
             if not dry_run:
-                done_task.status = schema.ARCHIVED_STATUS
+                archived_task.remove()
